array value in CLI and backup/restore fixes

+++ scripts/backup.ts
@@ -1,10 +1,9 @@
 import {FullStackedConfig} from "../index";
 import fs from "fs";
 import path from "path";
-import yaml from "yaml";
 import {execSync} from "child_process";
-import SFTP from "ssh2-sftp-client";
-import {execSSH} from "./utils";
+import {execSSH, getSFTPClient, getVolumesToBackup} from "./utils";
+import progress from "progress-stream";
 
 export default async function (config: FullStackedConfig) {
     if(config.host)
@@ -13,47 +12,74 @@ export default async function (config: FullStackedConfig) {
     if(!fs.existsSync(path.resolve(config.dist, "docker-compose.yml")))
         return console.log("Could not find built docker-compose file");
 
-    const dockerCompose = yaml.parse(fs.readFileSync(path.resolve(config.dist, "docker-compose.yml"), {encoding: "utf-8"}));
-    const volumes = Object.keys(dockerCompose.volumes);
+    const dockerComposeStr = fs.readFileSync(path.resolve(config.dist, "docker-compose.yml"), {encoding: "utf-8"});
+    const volumesToBackup = getVolumesToBackup(dockerComposeStr, config.volume);
 
-    if(!config.volume || !volumes.includes(config.volume))
-        return console.log("Volume not found in current docker-compose running");
+    for(const volume of volumesToBackup){
+        if(!config.silent)
+            console.log(`Backing up ${volume} from local host`);
 
-    execSync(`docker run -v ${config.name + "_" + config.volume}:/data -v ${config.backupDir ?? config.dist}/backup:/backup --name=fullstacked-backup busybox tar cvf backup/${config.volume}.tar data`, {
-        stdio: config.silent ? "ignore" : "inherit"
-    });
+        const commandArr = ["docker", "run",
+            "-v", config.name + "_" + volume + ":/data",
+            "-v", config.backupDir ?? path.resolve(process.cwd(), "backup") + ":/backup",
+            "--name=fullstacked-backup",
+            "busybox",
+            "tar cvf backup/" + volume + ".tar data"]
 
-    execSync(`docker rm fullstacked-backup -f -v`);
+        execSync(commandArr.join(" "), {
+            stdio: config.silent ? "ignore" : "inherit"
+        });
+
+        execSync(`docker rm fullstacked-backup -f -v`);
+    }
 }
 
 async function backupRemote(config: FullStackedConfig){
-    const sftp = new SFTP();
+    const sftp = await getSFTPClient(config);
 
-    // setup ssh connection
-    let connectionConfig: any = {
-        host: config.host,
-        username: config.user
-    }
+    const remoteDockerComposeFilePath = config.appDir + "/" + config.name + "/docker-compose.yml";
+    if(!await sftp.exists(remoteDockerComposeFilePath))
+        throw new Error("Cannot find docker-compose file in remote host");
 
-    if(config.sshPort)
-        connectionConfig.port = config.sshPort;
+    const dockerComposeBuffer = await sftp.get(remoteDockerComposeFilePath);
 
-    if(config.pass)
-        connectionConfig.password = config.pass;
+    const volumes = getVolumesToBackup(dockerComposeBuffer.toString(), config.volume);
 
-    if(config.privateKey)
-        connectionConfig.privateKey = fs.readFileSync(path.resolve(process.cwd(), config.privateKey));
-
-    await sftp.connect(connectionConfig);
-
-    await execSSH(sftp.client, `docker run -v ${config.name + "_" + config.volume}:/data -v /tmp/backup:/backup --name=fullstacked-backup busybox tar cvf backup/${config.volume}.tar data`);
-    await execSSH(sftp.client, `docker rm fullstacked-backup -f -v`);
-
-    const backupDir = path.resolve(config.backupDir ?? process.cwd(), "backup");
+    const backupDir = config.backupDir ?? path.resolve(process.cwd(), "backup");
     if(!fs.existsSync(backupDir))
         fs.mkdirSync(backupDir, {recursive: true});
 
-    await sftp.get(`/tmp/backup/${config.volume}.tar`, path.resolve(backupDir, `${config.volume}.tar`));
+    for(const volume of volumes){
+        if(!config.silent)
+            console.log(`Backing up ${volume} from remote host`);
+
+        const commandArr = ["docker", "run",
+            "-v", config.name + "_" + volume + ":/data",
+            "-v", "/tmp/backup:/backup",
+            "--name=fullstacked-backup",
+            "busybox",
+            "tar cvf backup/" + volume + ".tar data"
+        ]
+        await execSSH(sftp.client, commandArr.join(" "));
+        await execSSH(sftp.client, `docker rm fullstacked-backup -f -v`);
+
+        const tarFilePath = `/tmp/backup/${volume}.tar`;
+
+        const outFilePath = path.resolve(backupDir, `${volume}.tar`);
+        let dlStream = fs.createWriteStream(outFilePath);
+
+        if(!config.silent){
+            const fileStat = await sftp.stat(tarFilePath);
+
+            const progressStream = progress({
+                length: fileStat.size
+            });
+
+            dlStream = progressStream.pipe(dlStream);
+        }
+
+        await sftp.get(`/tmp/backup/${volume}.tar`, dlStream);
+    }
 
     // close connection
     await sftp.end();
